mingw-hdep: Add "maint set testuite-mode on/off" command
This patch is to add
maint set testsuite mode on/off
command.
  Finally, thanks to Yao's patch about handling of flushing of
stderr for mingw, it reduces to switching the
stdout and stderr between binary and text mode.

   Setting the file handles to binary mode avoid all the trouble
about double ^M that confuses DejaGnu so often when trying to
run the testsuite with a mingw compiled GDB.

Pierre Muller
GDB pascal language maintainer

2013-09-26  Pierre Muller  <muller@sourceware.org>

 	mingw-hdep.c (set_output_binary_mode): New function.
 	(set_output_normal_mode) : New function.
 	(maint_testsuite_mode): New static variable.
	(set_maint_testsuite_mode) : New function.
	(show_maint_testsuite_mode) : New function.
	(_initialize_mingw_hdep): Add "maint set testsuite-mode on/off"
 	command.

	(gdb_call_async_signal_handler):
	(void):