Fix handling of DW_AT_entry_pc of inlined subroutines
It may happen that the inline entry point is not the
start address of the first sub-range of an inline
function.

But the PC for a breakpoint on an inlined subroutine
is always the start address of the first sub-range.

This patch moves the sub-range starting at the entry
point to the first position of the block list.

Therefore the breakpoint on an inlined function
changes in rare cases from the start address of
the first sub-range to the real entry point.

There should always be a subrange that starts at the
entry point, even if that is an empty sub-range.

2021-01-01  Bernd Edlinger  <bernd.edlinger@hotmail.de>

	* dwarf2/read.c (dwarf2_record_block_ranges): Move range beginning
	at DW_AT_entry_pc to the first block range.