Return true in TuiWindow.is_valid only if TUI is enabled
There seems to be no other way to determine if TUI is enabled, which is a
problem in case the TUI is disabled, and you redraw the window contents
based on some registered event.
Then the TUI is redrawn, even though the TUI stays disabled, and the prompt
is anywhere on the screen.