Add optional first_column argument to TuiWindow.write
Horizontal scrolling of the window is not easy to implement if the text
contains ANSI terminal escape styling sequences.

So this new first_column argument of TuiWindow.write forwards the number
of columns, which should be skipped at the start of each line, to
tui_copy_source_line, since it already can take care of it via its first_col
argument.