Fix raw-frame-arguments in combination with frame-filters
Currently, if frame-filters are active, raw-values is used instead of
raw-frame-arguments to decide if a pretty-printer should be invoked for
frame arguments in a backtrace.

This adds the PRINT_RAW_FRAME_ARGUMENTS flag to frame_filter_flag which is
then used in the frame-filter to override the raw flag in enumerate_args.

gdb/ChangeLog:

2020-12-29  Hannes Domani  <ssbssa@yahoo.de>

	* extension.h (enum frame_filter_flag): Add
	PRINT_RAW_FRAME_ARGUMENTS.
	* python/py-framefilter.c (enumerate_args): Override raw flag.
	raw_frame_args argument.
	(py_mi_print_variables): Forward raw flag.
	(py_print_args): Forward raw_frame_args flag.
	(py_print_frame): Handle PRINT_RAW_FRAME_ARGUMENTS.
	* stack.c (backtrace_command_1): Set PRINT_RAW_FRAME_ARGUMENTS.

gdb/testsuite/ChangeLog:

2020-12-29  Hannes Domani  <ssbssa@yahoo.de>

	* gdb.python/py-frame-args.exp: Add bt raw-frame-arguments tests.
	* gdb.python/py-frame-args.py: Add basic frame-filter.