Coerce primitive types to target memory if reference is needed
It's currently not possible to call functions with reference
arguments, if you try to use a value that's not in the target
memory (like literal numbers):
```
(gdb) l f3
39      int f3(const int &i)
40      {
41        return i;
42      }
(gdb) p f3(10)
Attempt to take address of value not located in memory.
```

The reason for that is only arrays and strings are coerced to
target memory when using them as function call arguments.

To make sure a value is in target memory if a reference to it is
needed, this adds an extra try to coerce the value to target memory,
after casting the value to an array of 1 element.
Now this works:
```
(gdb) p f3(10)
$1 = 10
```

Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=15372
Bug: https://sourceware.org/bugzilla/show_bug.cgi?id=25957