gdb: restructure disable_breakpoints_in_unloaded_shlib
This commit rewrites disable_breakpoints_in_unloaded_shlib to be more
like disable_breakpoints_in_freed_objfile.  Instead of looping over
all b/p locations, we instead loop over all b/p and then over all
locations for each b/p.

The advantage of doing this is that we can fix this small bug that was
documented in a comment in the code:

  /* This may cause duplicate notifications for the same breakpoint.  */
  notify_breakpoint_modified (b);

By calling notify_breakpoint_modified() as we modify each location we
can potentially send multiple notifications for a single b/p.

Is this a bug?  Maybe not 100%.  After all, at each notification one
of the locations will have changed, so maybe its fine.  But it's not
ideal, and we can easily do better, so lets do that.

I originally made this change as part of a later commit in this
series, which fixed PR gdb/32079.  This rewrite isn't a requirement of
the later fix, but I do think this makes the function easier to
understand.

There's a new test which checks that we only get a single notification
when the shared library is unloaded.  Note that the test is written as
if there are multiple related but different tests within the same test
file ... but there aren't currently!  The next commit will add another
test proc to this test script at which point I think the comments will
make sense.  I've done this to avoid unnecessary churn in the next
commit.